Refactored helper code
